feat: MockApp react node for mock contexts
